working on world and network objects

+++ include/render_layer.h
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-enum class RenderLayer : uint_fast32_t
+enum class RenderLayer : uint_fast8_t
 {
     Opaque,
     Translucent,
